sudoku: Start pruning on "clumps"
If we have a case where there two cells in a group are both one of two
values, those two values cannot occur in any other cells in that , and
can be pruned from the list.

This can be generalised to any number of clumps (3 possibilites occuring
in 3 cells, etc.), but going past 3 probably won't help.